Fix some gcc compilation warnings:
- size_t vs. uint32_t
- operations mixing signed and unsigned
- Template the OpenGL Object m_handle type
- remove int operator because shall return size_t
- old C cast instead of static_cast
- missing destructors
- replacing m_ instead of _
- pedantic errors

+++ src/common/containers/CollectionIterator.ipp
@@ -143,5 +143,5 @@ protected:
   Collection<T, N, Block>* m_container;
 
   //! \brief the iterator.
-  uint32_t m_itr;
+  size_t m_itr;
 };
